Allow num0 to be any decimal number

+++ calc.py
@@ -4,7 +4,7 @@ def compute(expression):
     For example, '1 + 2' returns 3, and '1 - 2' returns -1.
     """
     num0, operator, num1 = expression.split(' ')
-    num0, num1 = int(num0), int(num1)
+    num0, num1 = float(num0), float(num1)
     if operator == '+':
         return num0 + num1
     elif operator == '-':
